+++ py/desisurvey/scripts/run_plan.py
@@ -17,7 +17,7 @@ def mjd_to_azstr(mjd):
 
 
 def run_plan(night=None, nts_dir=None, verbose=False, survey=None,
-             seeing=1.1):
+             seeing=1.1, table=False):
     kpno = EarthLocation.of_site('kpno')
     if nts_dir is None:
         obsplan = None
@@ -50,13 +50,16 @@ def run_plan(night=None, nts_dir=None, verbose=False, survey=None,
 
     print('local   lst   cond  tile    ra   dec    program   x fac  tot  split ' +
           'refft')
+    current_ra = None
+    current_dec = None
     while t0 < nts.scheduler.night_ephem['brightdawn']:
         cidx = np.interp(t0+300/86400, changes, np.arange(len(changes)))
         cidx = int(np.clip(cidx, 0, len(programs)-1))
         cond = programs[cidx]
         moon_up_factor = getattr(nts.config.conditions, cond).moon_up_factor()
         expdict = dict(mjd=t0, previoustiles=previoustiles)
-        conddict = dict(skylevel=moon_up_factor, seeing=seeing)
+        conddict = dict(skylevel=moon_up_factor, seeing=seeing,
+                        sky_ra=current_ra, sky_dec=current_dec)
         res = nts.next_tile(exposure=expdict, conditions=conddict,
                             speculative=True)
         if not res['foundtile']:
@@ -69,15 +72,32 @@ def run_plan(night=None, nts_dir=None, verbose=False, survey=None,
             nts.scheduler.tiles.tileID == res['fiberassign'])[0]
         ra = nts.scheduler.tiles.tileRA[ind]
         dec = nts.scheduler.tiles.tileDEC[ind]
+        current_ra = ra
+        current_dec = dec
         am = nts.scheduler.tiles.airmass_at_mjd(t0, mask=ind)
         refft = res['req_efftime']
-        print(
-            ('%s %5.1f %6s %5d %5.1f %5.1f %10s %3.1f %3.1f '
-             '%4d %6s %4d') % (
-                 mjd_to_azstr(t0), lst, res['conditions'],
-                 res['fiberassign'], ra, dec, res['program'],
-                 am, res['exposure_factor'], int(res['esttime']),
-                 ('%dx%d' % (res['count'], res['exptime'])), refft))
+        if not table:
+            print(
+                ('%s %5.1f %6s %5d %5.1f %5.1f %10s %3.1f %3.1f '
+                 '%4d %6s %4d') % (
+                     mjd_to_azstr(t0), lst, res['conditions'],
+                     res['fiberassign'], ra, dec, res['program'],
+                     am, res['exposure_factor'], int(res['esttime']),
+                     ('%dx%d' % (res['count'], res['exptime'])), refft))
+        else:
+            print(f"|| {mjd_to_azstr(t0)} || {lst:5.1f} "
+                  f"|| {res['fiberassign']} || "
+                  f"{res['ra']} || {res['dec']}"
+                  f"|| SV3 {res['program']:6s} || "
+                  f"{{{{{{sequence=DESI, exptime=5400.0, guider_exptime=5.0, "
+                  f"acquisition_exptime=15, "
+                  f"focus_exptime=60.0, sky_exptime=60.0, flavor=science, "
+                  f"useetc=True,program='{res['program']}',"
+                  f"fiberassign={res['fiberassign']},"
+                  f"req_efftime={res['req_efftime']:.0f}, "
+                  f"sbprof='{res['sbprof']}'"
+                  f"}}}}}} ||")
+
         t0 += (res['exptime']+0*180)*res['count']/60/60/24
 
 
@@ -96,6 +116,7 @@ def parse(options=None):
                         help='planning directory to use')
     parser.add_argument('--seeing', default=1.1, help='set seeing for night.',
                         type=float)
+    parser.add_argument('--table', default=False, action='store_true')
     if options is None:
         args = parser.parse_args()
     else:
@@ -105,4 +126,5 @@ def parse(options=None):
 
 def main(args):
     run_plan(night=args.night, nts_dir=args.nts_dir,
-             survey=args.survey, verbose=args.verbose, seeing=args.seeing)
+             survey=args.survey, verbose=args.verbose, seeing=args.seeing,
+             table=args.table)